[Refactor] : ViewModel 내부 폴더명 변경(페이지명 뒤에있는 model 글자 제거)
