--- app/src/main/java/kr/sparta/tripmate/data/model/ScrapModel.kt
@@ -1,4 +1,0 @@
-package kr.sparta.tripmate.data.model
-
-class ScrapModel(var title : String, var url : String, var description : String, var
-bloggername : String, var bloggerlink : String, var postdate : String)
