+++ app/src/main/java/kr/sparta/tripmate/ui/viewmodel/budget/BudgetFactory.kt
@@ -1,0 +1,4 @@
+package kr.sparta.tripmate.ui.viewmodel.budget
+
+class BudgetFactory {
+}
