+++ app/src/main/java/kr/sparta/tripmate/ui/viewmodel/budget/BudgetViewModel.kt
@@ -1,0 +1,4 @@
+package kr.sparta.tripmate.ui.viewmodel.budget
+
+class BudgetViewModel {
+}
