--- app/src/main/java/kr/sparta/tripmate/ui/viewmodel/budgetmodel/BudgetFactory.kt
@@ -1,4 +1,0 @@
-package kr.sparta.tripmate.ui.viewmodel.budgetmodel
-
-class BudgetFactory {
-}
