--- app/src/main/java/kr/sparta/tripmate/ui/viewmodel/budgetmodel/BudgetViewModel.kt
@@ -1,4 +1,0 @@
-package kr.sparta.tripmate.ui.viewmodel.budgetmodel
-
-class BudgetViewModel {
-}
