+++ app/src/main/java/kr/sparta/tripmate/ui/viewmodel/community/CommunityFactory.kt
@@ -1,0 +1,4 @@
+package kr.sparta.tripmate.ui.viewmodel.community
+
+class CommunityFactory {
+}
