--- app/src/main/java/kr/sparta/tripmate/ui/viewmodel/communitymodel/CommunityFactory.kt
@@ -1,4 +1,0 @@
-package kr.sparta.tripmate.ui.viewmodel.communitymodel
-
-class CommunityFactory {
-}
