+++ app/src/main/java/kr/sparta/tripmate/ui/viewmodel/home/HomeFactory.kt
@@ -1,0 +1,4 @@
+package kr.sparta.tripmate.ui.viewmodel.home
+
+class HomeFactory {
+}
