--- app/src/main/java/kr/sparta/tripmate/ui/viewmodel/homemodel/HomeFactory.kt
@@ -1,4 +1,0 @@
-package kr.sparta.tripmate.ui.viewmodel.homemodel
-
-class HomeFactory {
-}
