--- app/src/main/java/kr/sparta/tripmate/ui/viewmodel/homemodel/HomeViewModel.kt
@@ -1,4 +1,0 @@
-package kr.sparta.tripmate.ui.viewmodel.homemodel
-
-class HomeViewModel {
-}
