+++ app/src/main/java/kr/sparta/tripmate/ui/viewmodel/mypage/MyPageViewModel.kt
@@ -1,0 +1,4 @@
+package kr.sparta.tripmate.ui.viewmodel.mypage
+
+class MyPageViewModel {
+}
