--- app/src/main/java/kr/sparta/tripmate/ui/viewmodel/mypagemodel/MyPageFactory.kt
@@ -1,4 +1,0 @@
-package kr.sparta.tripmate.ui.viewmodel.mypagemodel
-
-class MyPageFactory {
-}
