--- app/src/main/java/kr/sparta/tripmate/ui/viewmodel/mypagemodel/MyPageViewModel.kt
@@ -1,4 +1,0 @@
-package kr.sparta.tripmate.ui.viewmodel.mypagemodel
-
-class MyPageViewModel {
-}
